refactor(rest): extract admin corpus routes into routes/admin_corpora.py
Moves the four admin corpus-management routes (GET/POST corpora,
PATCH/DELETE corpora/{slug}) out of create_app's closure into
routes/admin_corpora.py. Self-contained: only the DB pool via
require_pool, no new RestContext/deps.py surface.
